% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{treelite_test}
\alias{treelite_test}
\title{Test function}
\usage{
treelite_test(path)
}
\arguments{
\item{path}{Path to lightGBM model object}
}
\value{
Int
}
\description{
Test function
}
\examples{
treelite_test('LightGBM_model.txt')
}
